+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning].
 ### Changed in Unreleased
 
 - Keep pip in venv to allow Docker image to be used as a base image
+- Set venv ownership to user 1001 to allow pip install at runtime
 
 ## [4.1.0] - 2025-11-11