+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning].
 ### Changed in Unreleased
 
 - Keep pip in venv to allow Docker image to be used as a base image
-- Set venv ownership to user 1001 to allow pip install at runtime
+- Set venv permissions to allow any user to run pip install at runtime
+- Update test script to use szBuildVersion.json (renamed from g2BuildVersion.json)
 
 ## [4.1.0] - 2025-11-11